etl-processor-function/requirements.txt
azure-monitor-opentelemetry
azure-functions
azure-storage-blob
azure-servicebus
opentelemetry-api